test: add tests for Texas Holdem game engine's Ante round and state management
