+++ pvm/ts/src/engine/texasHoldem.ts
@@ -273,17 +273,14 @@ class TexasHoldemGame implements IPoker {
             throw new Error("Table full."); // This must be thrown
         }
 
-        // console.log(`Player ${player.address} joined at seat ${seat}`);
         this._playersMap.set(seat, player);
 
         // TODO: Need to consider the work flow here, but make active for now
         player.updateStatus(PlayerStatus.ACTIVE);
 
-        // if (player.chips < this._minBuyIn) {
-        //     // throw new Error("Player does not have enough chips to join.");
-        //     console.log("Player does not have enough chips to join.");
-        //     return;
-        // }
+        if (player.chips < this._gameOptions.minBuyIn) {
+            throw new Error("Player does not have enough chips to join.");
+        }
 
         const autoPostBlinds = false;
         if (autoPostBlinds) {
